bug fix: depth elec index number
elec_bin saved the depth elec according to the coord text file before
lh/rh splited, however if the input is lh or rh specific file, it will
be looking for wrong elecs for depth. Solution: separate the cortical
elecs and depth elecs by defining "hemi" to lh/rh/depth, as well as
making elec_bin as an optional input.  Also fix a bug that if the
orientation of elec_bin and aseg.nii.gz not the same, the anatomical
labels may be wrong. New codes check the orientation of elec_bin and
convert the aseg.mgz into the same orientation.